refactor: update imports to use Action type from domain/referentiels and fix typo in comment
