+++ app.territoiresentransitions.react/app/(authed)/collectivite/[collectiviteId]/referentiel/[referentielId]/(overview)/@tabs/progression/_components/axe.tsx
@@ -3,10 +3,9 @@ import { useSearchParams } from 'next/navigation';
 import { useEffect, useRef, useState } from 'react';
 
 import { ActionCard } from '@/app/referentiels/actions/action.card';
-import { Action as TActionBase } from '@/app/referentiels/actions/use-list-actions';
 import ScoreProgressBar from '@/app/referentiels/scores/score.progress-bar';
 import { ScoreRatioBadge } from '@/app/referentiels/scores/score.ratio-badge';
-import { ActionType } from '@/domain/referentiels';
+import { ActionType, Action as TActionBase } from '@/domain/referentiels';
 import { AccordionControlled, AccordionType } from '@/ui';
 
 export type TAxe = TActionBase & {
